refactors, work on signup
