+++ src/Cogitatio/Logic/Exceptions.cs
