+++ src/Cogitatio/Logic/SqlServer.cs
@@ -1,9 +1,10 @@
 ﻿using System.Data;
 using System.Transactions;
 using Cogitatio.Interfaces;
+using Cogitatio.Models;
 using Microsoft.Data.SqlClient;
 
-namespace Cogitatio.Models;
+namespace Cogitatio.Logic;
 
 /// <summary>
 /// TODO duplicity with Postgressql will be addressed in a future update
