+++ src/Cogitatio/Models/BlogUserRecord.cs
@@ -11,7 +11,8 @@ public class BlogUserRecord
     public string IpAddress { get; set; } = string.Empty;
     public string DisplayName { get; set; } = string.Empty;
     public UserAccountStates AccountState { get; set; } = UserAccountStates.Unknown;
-    public string TwoFactorSecret { get; set; }
+    public string TwoFactorSecret { get; set; } = string.Empty;
+    public string VerificationId { get; set; } = string.Empty;
     public string Password { get; set; }        // This password is always hashed
     public DateTime CreatedAt { get; set; }
