+++ src/schema/MSSQL/create_blog_users_tables.sql
@@ -1,12 +1,14 @@
 ﻿-- because we might want to keep user tables in a seaparate database
 -- the create statements are kept here for easy reference
 
+DROP TABLE IF EXISTS Blog_Users;
 CREATE TABLE Blog_Users (
     Id INT PRIMARY KEY IDENTITY(1,1),
     DisplayName NVARCHAR(50) NOT NULL UNIQUE,
     Email NVARCHAR(100) NOT NULL UNIQUE,
     IpAddress NVARCHAR(45) NOT NULL,                    -- 
     TwoFactorSecret NVARCHAR(50) NOT NULL,              -- generated by system
+    VerificationId NVARCHAR(50) NOT NULL,               -- generated by system, for email verification
     PasswordHash NVARCHAR(255) NOT NULL,                -- hashed password
     AccountState INT NOT NULL DEFAULT 0,                -- maps to enum UserAccountStates
     TenantId INT NOT NULL DEFAULT 0,                    -- comes from site settings
