+++ src/schema/Postgresql/create_blog_user_tables.sql
@@ -7,6 +7,7 @@ CREATE TABLE IF NOT EXISTS blog_users (
     email VARCHAR(100) NOT NULL UNIQUE,
     ip_address VARCHAR(45) NOT NULL,
     two_factor_secret VARCHAR(50) NOT NULL,
+    verification_id VARCHAR(50) NOT NULL,               -- generated by system, for email verification
     password_hash VARCHAR(255) NOT NULL,
     account_state INT NOT NULL DEFAULT 0,    -- maps to enum UserAccountStates
     tenant_id INT NOT NULL DEFAULT 0,
